typo on data type fix

+++ AspNetCoreVueJS/Models/Almacen/Articulo/ArticuloViewModel.cs
@@ -12,7 +12,7 @@ namespace AspNetCoreVueJS.Models.Almacen.Articulo
         public string categoria { get; set; }
         public string codigo { get; set; }
         public string nombre { get; set; }
-        public double precio_venta { get; set; }
+        public decimal precio_venta { get; set; }
         public int stock { get; set; }
         public string descripcion { get; set; }
         public bool condicion{ get; set; }
