+++ AspNetCoreVueJS/Models/Almacen/Articulo/UpdateProductViewModel.cs
@@ -16,7 +16,7 @@ namespace AspNetCoreVueJS.Models.Almacen.Articulo
         [StringLength(50, MinimumLength = 3, ErrorMessage = "This property has a max length of 50 and a minimum of 3")]
         public string nombre{ get; set; }
         [Required]
-        public double precio_venta{ get; set; }
+        public decimal precio_venta{ get; set; }
         [Required]
         public int stock { get; set; }
         public string descripcion{ get; set; }
