+++ NetCoreVueJSModels/Almacen/CArticulo.cs
@@ -14,7 +14,7 @@ namespace NetCoreVueJSModels.Almacen
         [StringLength(60, ErrorMessage ="El nombre tiene un máximo de 60 caracteres")]
         public string nombre { get; set; }
         [Required]
-        public double precio_venta { get; set; }
+        public decimal precio_venta { get; set; }
         public int stock { get; set; }
         public string descripcion { get; set; }
         public bool condicion { get; set; }
